size_t is not an int
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@9938 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/common/TextUtils.cxx
@@ -78,7 +78,7 @@ namespace TextUtils
 
   std::string no_whitespace(const std::string &s)
   {
-    const int sourcesize = s.size();
+    const int sourcesize = (int)s.size();
 
     int count = 0, i = 0, j = 0;
     for (i = 0; i < sourcesize; i++)
